requirements.txt
streamlit-pills
streamlit
SQLAlchemy
openai
llama-index==0.9.48
google-generativeai
transformers
torch
qdrant-client
fastembed
python-dotenv